refactor: Re-arrange components and update styles
