+++ frontend/src/components/BlockPositionHandler.vue
@@ -58,10 +58,9 @@
 	</div>
 </template>
 <script setup lang="ts">
+import OptionToggle from "@/components/Controls/OptionToggle.vue";
 import blockController from "@/utils/blockController";
 import { computed, watch } from "vue";
-import Input from "./Input.vue";
-import OptionToggle from "./OptionToggle.vue";
 
 const position = computed({
 	get() {
