+++ frontend/src/components/Controls/ImageUploader.vue
