+++ frontend/src/components/Controls/Switch.vue
@@ -1,0 +1,7 @@
+<template>
+	<Switch
+		class="[&>span>[id^=headlessui-description]]:text-text-icons-gray-9 [&>span>[id^=headlessui-label]]:text-text-icons-gray-6" />
+</template>
+<script setup lang="ts">
+import { Switch } from "frappe-ui";
+</script>
