+++ frontend/src/components/DimensionInput.vue
@@ -22,10 +22,10 @@
 		@update:modelValue="(val) => blockController.setStyle(property, val)"></InlineInput>
 </template>
 <script setup lang="ts">
+import InlineInput from "@/components/Controls/InlineInput.vue";
 import { styleProperty } from "@/utils/block";
 import blockController from "@/utils/blockController";
 import { PropType } from "vue";
-import InlineInput from "./InlineInput.vue";
 
 const props = defineProps({
 	property: {
