+++ frontend/src/components/Settings/GlobalGeneral.vue
@@ -47,14 +47,14 @@
 	</div>
 </template>
 <script setup lang="ts">
+import InlineInput from "@/components/Controls/InlineInput.vue";
 import { builderSettings } from "@/data/builderSettings";
 import { webPages } from "@/data/webPage";
 import useStore from "@/store";
 import { BuilderPage } from "@/types/Builder/BuilderPage";
 import { Switch } from "frappe-ui";
 import { computed } from "vue";
-import ImageUploader from "../ImageUploader.vue";
-import InlineInput from "../InlineInput.vue";
+import ImageUploader from "../Controls/ImageUploader.vue";
 
 const store = useStore();
